Revert "added telemetry properties"
This reverts commit b2477f8b48d3316a5dab2213fcae8ffb318d8e98.
