+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Xamls/PackageManagerControl.xaml.cs
@@ -624,7 +624,7 @@ namespace NuGet.PackageManagement.UI
             if (Model.Context.PackageRestoreManager != null)
             {
                 var projectContextInfo = Model.Context.Projects.Count() == 1 ? Model.Context.Projects.FirstOrDefault() : null;
-                RestoreBar = new PackageRestoreBar(Model.Context.SolutionManagerService, Model.Context.PackageRestoreManager, projectContextInfo, _detailModel.IsSolution);
+                RestoreBar = new PackageRestoreBar(Model.Context.SolutionManagerService, Model.Context.PackageRestoreManager, projectContextInfo);
                 DockPanel.SetDock(RestoreBar, Dock.Top);
 
                 _root.Children.Insert(0, RestoreBar);
