+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/NuGet.PackageManagement.VisualStudio.csproj
@@ -148,6 +148,7 @@
     <Compile Include="Telemetry\PackageManagerUIRefreshEvent.cs" />
     <Compile Include="Telemetry\DiagnosticTelemetryEvent.cs" />
     <Compile Include="Telemetry\ProjectTelemetryEvent.cs" />
+    <Compile Include="Telemetry\RestoreButtonAction.cs" />
     <Compile Include="Telemetry\SearchPageTelemetryEvent.cs" />
     <Compile Include="Telemetry\SearchSelectionTelemetryEvent.cs" />
     <Compile Include="Telemetry\SearchTelemetryEvent.cs" />
