+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/Telemetry/RestoreBannerClickedTelemetryEvent.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using NuGet.Common;
-using ContractsItemFilter = NuGet.VisualStudio.Internal.Contracts.ItemFilter;
 
 namespace NuGet.PackageManagement.Telemetry
 {
@@ -10,26 +9,11 @@ namespace NuGet.PackageManagement.Telemetry
     {
         internal const string RestoreBannerClickedEventName = "RestoreBannerClicked";
         internal const string RestoreButtonActionName = "RestoreButtonAction";
-        internal const string RestoreButtonOriginPropertyName = "IsSolutionView";
 
-        public RestoreBannerClickedTelemetryEvent(RestoreButtonAction restoreButtonAction, RestoreButtonOrigin restoreButtonOrigin)
+        public RestoreBannerClickedTelemetryEvent(RestoreButtonAction restoreButtonAction)
             : base(RestoreBannerClickedEventName)
         {
             base[RestoreButtonActionName] = restoreButtonAction;
-            base[RestoreButtonOriginPropertyName] = restoreButtonOrigin;
         }
-    }
-
-    public enum RestoreButtonAction
-    {
-        MissingAssetsFile,
-        MissingPackages
-    }
-
-    public enum RestoreButtonOrigin
-    {
-        SolutionView,
-        ProjectView,
-        PackageManagerConsole
     }
 }
