+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/Telemetry/RestoreButtonAction.cs
@@ -1,0 +1,11 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace NuGet.PackageManagement.Telemetry
+{
+    public enum RestoreButtonAction
+    {
+        MissingAssetsFile,
+        MissingPackages
+    }
+}
