+++ test/NuGet.Clients.Tests/NuGet.PackageManagement.VisualStudio.Test/NuGet.PackageManagement.VisualStudio.Test.csproj
@@ -53,7 +53,7 @@
     <Compile Include="Telemetry\HyperlinkClickedTelemetryEventTests.cs" />
     <Compile Include="Telemetry\IntervalTrackerTests.cs" />
     <Compile Include="Telemetry\NuGetTelemetryServiceTests.cs" />
-    <Compile Include="Telemetry\RestoreBannerClickedTelemetryEventTests.cs" />
+    <Compile Include="Telemetry\ResoreBannerClickedTelemetryEventTests.cs" />
     <Compile Include="Telemetry\RestoreTelemetryServiceTests.cs" />
     <Compile Include="Telemetry\SearchSelectionTelemetryEventTests.cs" />
     <Compile Include="Telemetry\TestTelemetryUtility.cs" />
